removed random move number (was too buggy)
