+++ UltimateGalaxyRandomizer/Randomizer/Galaxy.cs
@@ -87,7 +87,7 @@ namespace UltimateGalaxyRandomizer.Randomizer
                     player.Value.Skills[3].LearnAtLevel = 100;
                     break;
                 case 0xE8BF501E: //Arion Sherwind
-                    player.Value.Skills[0].SkillId = Moves.PlayerMoves.Where(x => x.Value.TP < 30 && x.Value.Type == MoveType.Dribble).Random().Key;
+                    player.Value.Skills[0].Skill = Moves.PlayerMoves.Where(x => x.Value.TP < 30 && x.Value.Type == MoveType.Dribble).Random().Value;
                     player.Value.Skills[0].SkillLevel = 1;
                     player.Value.Skills[0].LearnAtLevel = 0;
                     player.Value.Skills[3].LearnAtLevel = 100;
@@ -96,7 +96,7 @@ namespace UltimateGalaxyRandomizer.Randomizer
                     player.Value.Skills[3].LearnAtLevel = 100;
                     break;
                 case 0xFF7FE96D: //Victor Blade
-                    player.Value.Skills[0].SkillId = Moves.PlayerMoves.Where(x => x.Value.TP < 30 && x.Value.Type == MoveType.Shoot).Random().Key;
+                    player.Value.Skills[0].Skill = Moves.PlayerMoves.Where(x => x.Value.TP < 30 && x.Value.Type == MoveType.Shoot).Random().Value;
                     player.Value.Skills[0].SkillLevel = 1;
                     player.Value.Skills[0].LearnAtLevel = 0;
                     player.Value.Skills[3].LearnAtLevel = 100;
@@ -421,7 +421,7 @@ namespace UltimateGalaxyRandomizer.Randomizer
                     // Learn skill
                     if (player.Player.Skills[m].LearnAtLevel < team.Param.Level)
                     {
-                        player.Moves[m] = new SoccerMove(Moves.PlayerMoves[player.Player.Skills[m].SkillId], 1);
+                        player.Moves[m] = new SoccerMove(player.Player.Skills[m].Skill, 1);
                     } else
                     {
                         player.Moves[m] = null;
